Some last touches
I added a media quairy so the site will work for tablets and phones too. I also made sure to use the P element instead of h2 for the text in the article as h2 is for headers/ titles which is not what the poems in my article element is suposed to be.